♻️ [Refactor]: auth store 위치 및 이름 변경
