+++ src/entities/auth/store/auty.store.ts
